requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
idna==3.3
numpy==1.23.2
opencv-python==4.6.0.66
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
